Extract install application command
